refactor: make all packet class names end in "Packet"
