+++ sonar-common/src/main/java/xyz/jonesdev/sonar/common/fallback/protocol/FallbackPreparer.java
@@ -24,26 +24,26 @@ import xyz.jonesdev.sonar.common.fallback.protocol.block.BlockPosition;
 import xyz.jonesdev.sonar.common.fallback.protocol.block.BlockType;
 import xyz.jonesdev.sonar.common.fallback.protocol.block.ChangedBlock;
 import xyz.jonesdev.sonar.common.fallback.protocol.map.MapInfoPreparer;
-import xyz.jonesdev.sonar.common.fallback.protocol.packets.config.FinishConfiguration;
-import xyz.jonesdev.sonar.common.fallback.protocol.packets.config.RegistryData;
+import xyz.jonesdev.sonar.common.fallback.protocol.packets.config.FinishConfigurationPacket;
+import xyz.jonesdev.sonar.common.fallback.protocol.packets.config.RegistryDataPacket;
 import xyz.jonesdev.sonar.common.fallback.protocol.packets.play.*;
 
 @UtilityClass
 public class FallbackPreparer {
 
   // Abilities
-  public final FallbackPacket DEFAULT_ABILITIES = new Abilities(0x00, 0f, 0f);
-  public final FallbackPacket CAPTCHA_ABILITIES = new Abilities(0x02, 0f, 0f);
+  public final FallbackPacket DEFAULT_ABILITIES = new ClientAbilitiesPacket(0x00, 0f, 0f);
+  public final FallbackPacket CAPTCHA_ABILITIES = new ClientAbilitiesPacket(0x02, 0f, 0f);
   // Chunks
-  public final FallbackPacket EMPTY_CHUNK_DATA = new EmptyChunkData(0, 0);
+  public final FallbackPacket EMPTY_CHUNK_DATA = new ChunkDataPacket(0, 0);
   // Finish Configuration
-  public final FallbackPacket FINISH_CONFIGURATION = new FinishConfiguration();
+  public final FallbackPacket FINISH_CONFIGURATION = new FinishConfigurationPacket();
   // Synchronize Registry
-  public final FallbackPacket REGISTRY_SYNC = new RegistryData();
+  public final FallbackPacket REGISTRY_SYNC = new RegistryDataPacket();
   // Keep Alive
-  public final FallbackPacket CAPTCHA_KEEP_ALIVE = new KeepAlive(0L);
+  public final FallbackPacket CAPTCHA_KEEP_ALIVE = new KeepAlivePacket(0L);
   // Game Event (1.20.3+)
-  public final FallbackPacket START_WRITING_CHUNKS = new GameEvent(13, 0);
+  public final FallbackPacket START_WRITING_CHUNKS = new GameEventPacket(13, 0);
   // Chat
   public FallbackPacket enterCodeMessage;
   public FallbackPacket incorrectCaptcha;
@@ -69,9 +69,9 @@ public class FallbackPreparer {
   public final int DEFAULT_Y_COLLIDE_POSITION = 255; // 255 is the maximum Y position allowed
 
   // Captcha position
-  public final FallbackPacket CAPTCHA_POSITION = new PositionLook(
+  public final FallbackPacket CAPTCHA_POSITION = new PlayerPositionLookPacket(
     SPAWN_X_POSITION, 1337, SPAWN_Z_POSITION, 0f, 90f, 0, false);
-  public final FallbackPacket CAPTCHA_SPAWN_POSITION = new DefaultSpawnPosition(
+  public final FallbackPacket CAPTCHA_SPAWN_POSITION = new DefaultSpawnPositionPacket(
     SPAWN_X_POSITION, 1337, SPAWN_Z_POSITION);
 
   // Blocks
@@ -83,7 +83,7 @@ public class FallbackPreparer {
   public double maxFallDistance;
 
   public void prepare() {
-    joinGame = new JoinGame(0,
+    joinGame = new JoinGamePacket(0,
       Sonar.get().getConfig().getVerification().getGravity().getGamemode().getId(),
       0,
       false,
@@ -106,7 +106,7 @@ public class FallbackPreparer {
 
     // Set the dynamic block and collide Y position based on the maximum fall distance
     dynamicSpawnYPosition = DEFAULT_Y_COLLIDE_POSITION + (int) Math.ceil(maxFallDistance);
-    dynamicSpawnPosition = new DefaultSpawnPosition(SPAWN_X_POSITION, dynamicSpawnYPosition, SPAWN_Z_POSITION);
+    dynamicSpawnPosition = new DefaultSpawnPositionPacket(SPAWN_X_POSITION, dynamicSpawnYPosition, SPAWN_Z_POSITION);
 
     // Prepare collision platform positions
     int index = 0;
@@ -122,21 +122,21 @@ public class FallbackPreparer {
     }
 
     // Prepare UpdateSectionBlocks packet
-    updateSectionBlocks = new UpdateSectionBlocks(0, 0, CHANGED_BLOCKS);
+    updateSectionBlocks = new UpdateSectionBlocksPacket(0, 0, CHANGED_BLOCKS);
 
     // Prepare disconnect packets during login
-    blacklisted = Disconnect.create(Sonar.get().getConfig().getVerification().getBlacklisted(), true);
-    alreadyVerifying = Disconnect.create(Sonar.get().getConfig().getVerification().getAlreadyVerifying(), true);
-    alreadyQueued = Disconnect.create(Sonar.get().getConfig().getVerification().getAlreadyQueued(), true);
-    protocolBlacklisted = Disconnect.create(Sonar.get().getConfig().getVerification().getProtocolBlacklisted(), true);
-    reconnectedTooFast = Disconnect.create(Sonar.get().getConfig().getVerification().getTooFastReconnect(), true);
-    invalidUsername = Disconnect.create(Sonar.get().getConfig().getVerification().getInvalidUsername(), true);
+    blacklisted = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getBlacklisted(), true);
+    alreadyVerifying = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getAlreadyVerifying(), true);
+    alreadyQueued = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getAlreadyQueued(), true);
+    protocolBlacklisted = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getProtocolBlacklisted(), true);
+    reconnectedTooFast = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getTooFastReconnect(), true);
+    invalidUsername = DisconnectPacket.create(Sonar.get().getConfig().getVerification().getInvalidUsername(), true);
 
     // Precompute captcha answers
     if (Sonar.get().getConfig().getVerification().getMap().getTiming() != SonarConfiguration.Verification.Timing.NEVER) {
-      enterCodeMessage = new Chat(Sonar.get().getConfig().getVerification().getMap().getEnterCode(), Chat.SYSTEM_TYPE);
-      incorrectCaptcha = new Chat(
-        Sonar.get().getConfig().getVerification().getMap().getFailedCaptcha(), Chat.SYSTEM_TYPE);
+      enterCodeMessage = new UniversalChatPacket(Sonar.get().getConfig().getVerification().getMap().getEnterCode(), UniversalChatPacket.SYSTEM_TYPE);
+      incorrectCaptcha = new UniversalChatPacket(
+        Sonar.get().getConfig().getVerification().getMap().getFailedCaptcha(), UniversalChatPacket.SYSTEM_TYPE);
 
       MapInfoPreparer.prepare();
     }
